Fixed misparse of fdisk output.
When a partition has an odd number of sectors, the size field printed
by fdisk is "<size in KiB>+", not "<size in KiB>".

This causes a syntax error when the extracted field is pasted into
an expression to evaluate.

The fix is to subtract the start sector index from end end sector index
(as is already done for the linux partition).