This is a temporary workaround for bug #816015.
Prevent the indexing of any snapshots that aren't in http://snapshots.linaro.org/11.05-daily/ so the current logic for hwpack and OS selection will not result in finding an OS binary without a hardware pack to go with it.

When testing I noticed that I hadn't changed the name of the library FetchImage to fetch_image (without this change the indexer wouldn't work). I also moved the BeautifulSoup import in fetch_image.py to the single function where it is used, so BeautifulSoup is not needed to be installed on the server that is creating the index.