Proper handling of the new way of telling the indexer to index a directory:
crawler.add_directory_parse_list(
          RELEASES_WWW_DOCUMENT_ROOT,
          RELEASE_URL,

          (["hwpack"], ["/platform/"]),
This tells the indexer to look for a URL that matches the string "hwpack", but doesn't match "/platform/".

          ["platform", "build", "hardware"],
These are the columns to put in the database

          "release_hwpacks",
          ["platform", "", "", ("hardware", r"hwpack_linaro-(.*?)_")])
Once a URL has been split up by os.sep, this is how to file what is found. In this instance the first directory is used as the platform value in the database entry to be created. The tuple is of the form (column name, regexp), where regexp.group(1) is used as the value stored in the database in the colum "column name". In this example the build column is not filled in, but it is created in the database. This is to allow for a database where sometimes this column is used. This will be the case when transitioning from the old to new server layout.

The UI currently won't cope with the NULL entry in the build column properly, but at least the indexer will work.

This has been tested with an example layout of the new server structure, but not with the old server structure, so is not ready to merge.