ODROID: add to install secured boot loader
Exynos4412 must use secured boot loader signed by Hardkernel, it is consisted
with 4 binaries as BL1.bin, BL2.bin, u-boot.bin and TZSW.bin. And this patch
is to flash the files at specific sector on the boot device.

These files are included in Hardkernel's BSP release as a binary form and
packaged as 'u-boot-odroid' at Launchpad. This package will be downloaded and
installed during 'linaro-media-create' runs and flashed.

Signed-off-by: Dongjin Kim <tobetter@gmail.com>