Add a generic board config
Add a generic config for boards that drop a kernel and an initrd into
the boot partition.

Juno is such a board, where the other boot files, such as DTB and
firmware are put into the boot partition by the hardware pack metadata.

FastModel config now herits from the generic config.

Change-Id: Ib60c57f738eaa0199bbefd49cf77e0cd4a9ddb97
Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>
Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>