unpack_binary_tarball.py: add --selinux option for tar command
to support selinux for android with tarball files,
we need to use the --selinux option of tar command
to retain the security label information to filesystem from tarball files.
and here we pass the --selinux option only when the tar command in the OS supports it

except the --selinux option for tar command, changed the method
to populate content for system/userdata partition.
since with the move method, it can not keep the security information
for the root directory of that partition.

Change-Id: I1beef41042a883f31abd2a658c8368a15fa08247
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>